Create a Model trait
