+++ llama-rs/Cargo.toml
@@ -9,18 +9,10 @@ rust-version = "1.65"
 [dependencies]
 ggml = { path = "../ggml" }
 
+bytemuck = { workspace = true }
 rand = { workspace = true }
+serde = { workspace = true }
 
-bytemuck = "1.13.1"
-partial_sort = "0.2.0"
 thiserror = "1.0"
-serde = { version = "1.0", features = ["derive"] }
+partial_sort = "0.2.0"
 serde_bytes = "0.11"
-
-# Used for the `convert` feature
-serde_json = { version = "1.0", optional = true }
-protobuf = { version = "= 2.14.0", optional = true }
-rust_tokenizers = { version = "3.1.2", optional = true }
-
-[features]
-convert = ["dep:serde_json", "dep:protobuf", "dep:rust_tokenizers"]
