+++ llama-rs/src/inference_session.rs
@@ -5,7 +5,7 @@ use rand::{distributions::WeightedIndex, prelude::Distribution};
 use thiserror::Error;
 
 use crate::{
-    util::mulf, EvaluateOutputRequest, InferenceError, InferenceParameters, Model, TokenId,
+    mulf, EvaluateOutputRequest, InferenceError, InferenceParameters, Model, TokenId,
     TokenUtf8Buffer, EOT_TOKEN_ID,
 };
 
@@ -28,33 +28,36 @@ pub struct InferenceSession {
     // Parameters for the session.
     pub(crate) params: InferenceSessionParameters,
 
-    pub(crate) memory_k: ggml::Tensor,
-    pub(crate) memory_v: ggml::Tensor,
+    /// Memory K
+    pub memory_k: ggml::Tensor,
+
+    /// Memory M
+    pub memory_v: ggml::Tensor,
 
     /// How many tokens have been fed into the model's working memory so far.
-    pub(crate) n_past: usize,
+    pub n_past: usize,
 
     /// How much memory is required per token for the temporary context used
     /// during inference.
-    pub(crate) mem_per_token: usize,
+    pub mem_per_token: usize,
 
     /// All tokens generated by this inference session
     pub(crate) tokens: Vec<TokenId>,
 
     /// The logits that were last predicted by the network. Zeroed out otherwise.
-    pub(crate) last_logits: Vec<f32>,
+    pub last_logits: Vec<f32>,
 
     /// Scratch buffers used during inference.
     ///
     /// The number of scratch buffers was copied from `llama.cpp`.
     /// There is no specific reason for this number, but one is insufficient.
-    pub(crate) scratch: [ggml::Buffer; 2],
+    pub scratch: [ggml::Buffer; 2],
 }
 impl InferenceSession {
     /// Feed a prompt to the model for this session.
     pub fn feed_prompt<E: std::error::Error + 'static>(
         &mut self,
-        model: &Model,
+        model: &impl Model,
         params: &InferenceParameters,
         prompt: &str,
         mut callback: impl FnMut(&[u8]) -> Result<(), E>,
@@ -68,7 +71,7 @@ impl InferenceSession {
             .map(|(_, tok)| *tok)
             .collect();
 
-        if self.n_past + prompt_tokens.len() >= model.hparams.n_ctx {
+        if self.n_past + prompt_tokens.len() >= model.n_ctx() {
             return Err(InferenceError::ContextFull);
         }
 
@@ -92,11 +95,11 @@ impl InferenceSession {
     /// Infer the next token for this session.
     pub fn infer_next_token<'v>(
         &mut self,
-        model: &'v Model,
+        model: &'v impl Model,
         params: &InferenceParameters,
         rng: &mut impl rand::Rng,
     ) -> Result<&'v [u8], InferenceError> {
-        if self.n_past + 1 >= model.hparams.n_ctx {
+        if self.n_past + 1 >= model.n_ctx() {
             return Err(InferenceError::ContextFull);
         }
 
@@ -129,7 +132,7 @@ impl InferenceSession {
     /// If `params.play_back_previous_tokens` is specified, this will "play back" all existing tokens in the session.
     pub fn inference_with_prompt<E: std::error::Error + 'static>(
         &mut self,
-        model: &Model,
+        model: &impl Model,
         params: &InferenceParameters,
         prompt: &str,
         maximum_token_count: Option<usize>,
@@ -311,7 +314,7 @@ impl InferenceSession {
     /// Creates an [InferenceSession] from a snapshot.
     pub fn from_snapshot(
         snapshot: InferenceSnapshot,
-        model: &Model,
+        model: &impl Model,
     ) -> Result<Self, SnapshotError> {
         let mut session = model.start_session(snapshot.session_params);
 
@@ -340,7 +343,8 @@ impl InferenceSession {
     }
 }
 impl InferenceSession {
-    pub(crate) fn new(
+    /// Create a new InferenceSession
+    pub fn new(
         params: InferenceSessionParameters,
         n_ctx: usize,
         n_layer: usize,
