+++ llama/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "llama"
+version = { workspace = true }
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bytemuck = { workspace = true }
+ggml = { path = "../ggml" }
+llama-rs = { path = "../llama-rs" }
+
+serde = { workspace = true }
+
+thiserror = "1.0"
+
+# Used for the `convert` feature
+serde_json = { version = "1.0", optional = true }
+protobuf = { version = "= 2.14.0", optional = true }
+rust_tokenizers = { version = "3.1.2", optional = true }
+
+[features]
+convert = ["dep:serde_json", "dep:protobuf", "dep:rust_tokenizers"]
